pyproject.toml
[project]
name = "datamission1"
version = "0.1.0"
description = ""
authors = [
    {name = "Nicolas",email = "nicolascnagel@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas (>=3.0.2,<4.0.0)",
    "requests (>=2.33.1,<3.0.0)",
    "fastapi (>=0.136.1,<0.137.0)",
    "uvicorn (>=0.46.0,<0.47.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "psycopg2-binary (>=2.9.12,<3.0.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
